# the-api-swagger.yaml
openapi: 3.0.3
info:
  title: THE database REST API
  description: |
    A REST API to fetch information from the THE database.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
externalDocs:
  description: Our Jira project
  url: https://apps.nrs.gov.bc.ca/int/jira/projects/FSADT2
servers:
  - url: https://nrbestapi-test-service-api.apps.silver.devops.gov.bc.ca/api
tags:
  - name: vegetationCode
    description: Codes describing various vegetation species
paths:
  /vegetationCode:
    get:
      tags:
        - vegetationCode
      summary: Search for valid vegetation codes by their identifier or description
      description: |
        Search for valid vegetation codes (ones which the interval between `effectiveDate` and
        `expiryDate` includes the present) with identifier or description that match `search`.
      operationId: searchVegetationCodes
      parameters:
        - name: search
          in: query
          description: |
            A string to which the code's identifier or description should match. Default value is an
            empty string, which matches anything.
          required: false
          schema:
            type: string
        - schema:
          $ref: '#/components/parameters/Page'
        - schema:
          $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: An array with the vegetation codes found, ordered by their identifiers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VegetationCode'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /vegetationCode/{code}:
    get:
      tags:
        - vegetationCode
      summary: Fetch a vegetation code by its identifier
      description: |
        Returns a single vegetation code identified by `code`.
      operationId: getVegetationCodeById
      parameters:
        - name: code
          in: path
          description: identifier of the vegetation code being sought
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VegetationCode'
        '404':
          description: Vegetation code not found.
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    VegetationCode:
      type: object
      properties:
        id:
          type: string
          maxLength: 8
          description: |
            The code itself.
        description:
          type: string
          maxLength: 120
          description: |
            Description for the affiliated code.
        effectiveDate:
          type: string
          format: date
          description: |
            The date from which the code is in effect.
        expiryDate:
          type: string
          format: date
          description: |
            The date on which the code expires.
  parameters:
    Page:
      name: page
      in: query
      description: |
        The zero-indexed number of the page to display; must be non-negative. Defaults to 0.
      schema:
        type: integer
    PageSize:
      name: pageSize
      in: query
      description: |
        The maximum number of results to fetch; must be positive. Defaults to 20.
      schema:
        type: integer
  responses:
    InternalServerError:
      description: |
        Something else went wrong while processing the request.
    BadRequest:
      description: |
        Parameters' constraints were not met.